requirements.txt
# Core dependencies - Modelos Locais Sem Censura
transformers>=4.41.0
peft>=0.10.0
trl>=0.8.0
datasets>=2.18.0
qdrant-client>=1.9.0
sentence-transformers>=2.6.0
tree-sitter>=0.20.2
tree-sitter-python>=0.20.4
tree-sitter-javascript>=0.20.3
tree-sitter-typescript>=0.20.3
faiss-cpu>=1.8.0
docker>=7.0.0
typer>=0.12.0
pydantic>=2.7.0
rich>=13.7.0
pytest>=8.1.0
coverage>=7.5.0
ruff>=0.4.0
mypy>=1.10.0
python-dotenv>=1.0.1
bitsandbytes; sys_platform == "linux"
torch>=2.0.0
accelerate>=0.20.0
scipy>=1.10.0
numpy>=1.24.0
requests>=2.31.0
pyyaml>=6.0.0

# Modelos Locais - DeepSeek e outros sem censura
ollama>=0.2.0
llama-cpp-python>=0.2.0
vllm>=0.4.0
auto-gptq>=0.7.0
optimum>=1.18.0
ctransformers>=0.2.0

# API and web dependencies
fastapi>=0.111.0
uvicorn>=0.30.0
sse-starlette>=2.0.0
httpx>=0.27.0
pinecone-client>=3.0.0
uvloop>=0.19.0
httptools>=0.6.0
